+++ packages/components/src/components/Image/index.ts
@@ -3,7 +3,6 @@ import {
   CONTENTFUL_COMPONENTS,
   CONTENTFUL_DEFAULT_CATEGORY,
 } from '@contentful/experience-builder-core/constants';
-import { constants } from '@/utils/constants';
 
 export * from './Image';
 
@@ -23,7 +22,6 @@ export const ImageComponentDefinition: ComponentDefinition = {
     'cfImageQuality',
     'cfImageSizes',
   ],
-  thumbnailUrl: constants.thumbnails.image,
   variables: {
     alt: {
       displayName: 'Alt',